Refactor category input to dropdown in Edit view
Updated StoresController.cs to include a predefined list of categories ("Electronics", "Clothing", "Books", "Groceries") and assigned it to ViewBag.CategoryList. Modified Edit.cshtml to use this list to populate a dropdown menu for the "Category" field, replacing the previous text input. Added a default option "-- Selecteer Categorie --" to the dropdown.